Feature 나스닥 정보 홈뷰 생성

+++ client/src/app/utils/formatters.ts
@@ -2,7 +2,7 @@ export function formatPrice(price: string, exchangeCode: string) {
   if (!price) return '0원';
   const numericPrice = price.replace(/[^\d.-]/g, '');
   if (exchangeCode !== 'KRX') {
-    return `${parseInt(numericPrice).toLocaleString()}$`;
+    return `${parseFloat(numericPrice).toLocaleString()}$`;
   }
   return `${parseInt(numericPrice).toLocaleString()}원`;
 }
